requirements.txt
bs4==0.0.1
numpy==1.21.0
pandas==1.3
pre-commit
python-dotenv==0.19.0
requests==2.26.0
selenium==3.141.0
XlsxWriter==1.4.5
